+++ pkg/workflows/management/create.go
@@ -4,16 +4,14 @@ import (
 	"context"
 
 	"github.com/aws/eks-anywhere/pkg/cluster"
-	"github.com/aws/eks-anywhere/pkg/constants"
 	"github.com/aws/eks-anywhere/pkg/filewriter"
 	"github.com/aws/eks-anywhere/pkg/providers"
 	"github.com/aws/eks-anywhere/pkg/task"
-	"github.com/aws/eks-anywhere/pkg/types"
 	"github.com/aws/eks-anywhere/pkg/workflows/interfaces"
 )
 
-// CreateManagement is a schema for create cluster.
-type CreateManagement struct {
+// Create is a schema for create cluster.
+type Create struct {
 	bootstrapper     interfaces.Bootstrapper
 	provider         providers.Provider
 	clusterManager   interfaces.ClusterManager
@@ -24,14 +22,14 @@ type CreateManagement struct {
 	clusterUpgrader  interfaces.ClusterUpgrader
 }
 
-// NewCreateManagement builds a new create construct.
-func NewCreateManagement(bootstrapper interfaces.Bootstrapper, provider providers.Provider,
+// NewCreate builds a new create construct.
+func NewCreate(bootstrapper interfaces.Bootstrapper, provider providers.Provider,
 	clusterManager interfaces.ClusterManager, gitOpsManager interfaces.GitOpsManager,
 	writer filewriter.FileWriter, eksdInstaller interfaces.EksdInstaller,
 	packageInstaller interfaces.PackageInstaller,
 	clusterUpgrade interfaces.ClusterUpgrader,
-) *CreateManagement {
-	return &CreateManagement{
+) *Create {
+	return &Create{
 		bootstrapper:     bootstrapper,
 		provider:         provider,
 		clusterManager:   clusterManager,
@@ -44,14 +42,7 @@ func NewCreateManagement(bootstrapper interfaces.Bootstrapper, provider provider
 }
 
 // Run runs all the create management cluster tasks.
-func (c *CreateManagement) Run(ctx context.Context, clusterSpec *cluster.Spec, validator interfaces.Validator, forceCleanup bool) error {
-	if forceCleanup {
-		if err := c.bootstrapper.DeleteBootstrapCluster(ctx, &types.Cluster{
-			Name: clusterSpec.Cluster.Name,
-		}, constants.Create, forceCleanup); err != nil {
-			return err
-		}
-	}
+func (c *Create) Run(ctx context.Context, clusterSpec *cluster.Spec, validator interfaces.Validator, forceCleanup bool) error {
 	commandContext := &task.CommandContext{
 		Bootstrapper:     c.bootstrapper,
 		Provider:         c.provider,
@@ -65,11 +56,5 @@ func (c *CreateManagement) Run(ctx context.Context, clusterSpec *cluster.Spec, v
 		ClusterUpgrader:  c.clusterUpgrader,
 	}
 
-	if clusterSpec.ManagementCluster != nil {
-		commandContext.BootstrapCluster = clusterSpec.ManagementCluster
-	}
-
-	err := task.NewTaskRunner(&setupAndValidateCreate{}, c.writer).RunTask(ctx, commandContext)
-
-	return err
+	return task.NewTaskRunner(&setupAndValidateCreate{}, c.writer).RunTask(ctx, commandContext)
 }
